added hibernate JPA.

+++ src/main/java/com/mk/Dao/UserDao.java
@@ -6,14 +6,14 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 
 import javax.annotation.Resource;
+import java.sql.Date;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.Date;
 import java.util.List;
 import java.util.UUID;
 
 @Repository
-public class UserDao {
+public class UserDao{
 
     @Resource
     private JdbcTemplate jdbcTemplate;
@@ -29,7 +29,7 @@ public class UserDao {
                 user.setFirst_name(rs.getString("first_name"));
                 user.setLast_name(rs.getString("last_name"));
                 user.setGender(rs.getString("gender"));
-                user.setDate_of_birth(rs.getString("date_of_birth"));
+                user.setDate_of_birth(Date.valueOf(rs.getString("date_of_birth")));
                 user.setEmail(rs.getString("email"));
                 return user;
             }
@@ -48,7 +48,7 @@ public class UserDao {
                 user.setFirst_name(rs.getString("first_name"));
                 user.setLast_name(rs.getString("last_name"));
                 user.setGender(rs.getString("gender"));
-                user.setDate_of_birth(rs.getString("date_of_birth"));
+                user.setDate_of_birth(Date.valueOf(rs.getString("date_of_birth")));
                 user.setEmail(rs.getString("email"));
                 return user;
             }
@@ -62,14 +62,26 @@ public class UserDao {
         /*
         need to convert String to Date type so the database can recognize.
          */
-        jdbcTemplate.update(sql,java.util.UUID.fromString(user.getId()),user.getFirst_name(),user.getLast_name(),user.getGender(),java.sql.Date.valueOf(user.getDate_of_birth()),user.getEmail());
+        jdbcTemplate.update(sql,
+                user.getId(),
+                user.getFirst_name(),
+                user.getLast_name(),
+                user.getGender(),
+                user.getDate_of_birth(),
+                user.getEmail());
         return "User "+user.getFirst_name()+" "+user.getLast_name()+" added";
 
     }
 
-    public String removeUser(String id){
+    public String removeUser(String id) throws SQLException {
+        if (!verifyId(id)) throw new SQLException();
         String sql = "DELETE FROM users WHERE id = ?";
-        jdbcTemplate.update(sql,java.util.UUID.fromString(id));
+        jdbcTemplate.update(sql,id);
         return "User "+id+" removed";
+    }
+
+    private boolean verifyId(String id){
+        String sql = "SELECT EXISTS(SELECT FROM users WHERE id = ?)";
+        return jdbcTemplate.queryForObject(sql, Boolean.class, id);
     }
 }
